+++ changelog.md
@@ -10,6 +10,7 @@
 - fix bug in sync_client @maramalshen
 - add detail with the action.sync_id
 - if create transfer was failed then send command: Adjust Inventory (optionals)
+- [join] if repzo.joinActionsWebHook.status = failure then commandLog.setStatus should be fail @maramalshen
 
 ### Changed